Add swipe to delete

+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/vaccination/ui/list/VaccinationListFragment.kt
@@ -10,13 +10,15 @@ import androidx.navigation.fragment.FragmentNavigatorExtras
 import androidx.navigation.fragment.findNavController
 import androidx.navigation.fragment.navArgs
 import com.google.android.material.appbar.AppBarLayout
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.google.android.material.imageview.ShapeableImageView
 import de.rki.coronawarnapp.R
 import de.rki.coronawarnapp.databinding.FragmentVaccinationListBinding
 import de.rki.coronawarnapp.ui.qrcode.fullscreen.QrCodeFullScreenFragmentArgs
 import de.rki.coronawarnapp.ui.view.onOffsetChange
-import de.rki.coronawarnapp.util.DialogHelper
 import de.rki.coronawarnapp.util.di.AutoInject
+import de.rki.coronawarnapp.util.list.isSwipeable
+import de.rki.coronawarnapp.util.list.onSwipeItem
 import de.rki.coronawarnapp.util.lists.diffutil.update
 import de.rki.coronawarnapp.util.ui.doNavigate
 import de.rki.coronawarnapp.util.ui.popBackStack
@@ -47,7 +49,7 @@ class VaccinationListFragment : Fragment(R.layout.fragment_vaccination_list), Au
         }
     )
 
-    private val adapter = VaccinationListAdapter()
+    private val vaccinationListAdapter = VaccinationListAdapter()
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
@@ -57,7 +59,15 @@ class VaccinationListFragment : Fragment(R.layout.fragment_vaccination_list), Au
                 popBackStack()
             }
 
-            recyclerViewVaccinationList.adapter = adapter
+            recyclerViewVaccinationList.apply {
+                adapter = vaccinationListAdapter
+                onSwipeItem(requireContext()) { position, direction ->
+                    val vaccinationItem = vaccinationListAdapter.data[position]
+                    if (vaccinationItem.isSwipeable()) {
+                        vaccinationItem.onSwipe(position, direction)
+                    }
+                }
+            }
 
             viewModel.uiState.observe(viewLifecycleOwner) { uiState ->
                 bindViews(uiState)
@@ -87,7 +97,7 @@ class VaccinationListFragment : Fragment(R.layout.fragment_vaccination_list), Au
                         )
                     }
                     is DeleteVaccinationEvent -> {
-                        showDeleteVaccinationDialog(event.vaccinationCertificateId)
+                        showDeleteVaccinationDialog(event.vaccinationCertificateId, event.position)
                     }
                     is NavigateBack -> {
                         popBackStack()
@@ -107,7 +117,7 @@ class VaccinationListFragment : Fragment(R.layout.fragment_vaccination_list), Au
 
         setToolbarOverlay()
 
-        adapter.update(listItems)
+        vaccinationListAdapter.update(listItems)
 
         val background = if (hasImmunity) {
             R.drawable.vaccination_compelete_gradient
@@ -141,19 +151,20 @@ class VaccinationListFragment : Fragment(R.layout.fragment_vaccination_list), Au
         behavior.overlayTop = (deviceWidth / divider) - 24
     }
 
-    private fun showDeleteVaccinationDialog(vaccinationCertificateId: String) {
-        DialogHelper.showDialog(
-            DialogHelper.DialogInstance(
-                requireActivity(),
-                R.string.vaccination_list_deletion_dialog_title,
-                R.string.vaccination_list_deletion_dialog_message,
-                R.string.vaccination_list_deletion_dialog_positive_button,
-                R.string.vaccination_list_deletion_dialog_negative_button,
-                positiveButtonFunction = {
-                    viewModel.deleteVaccination(vaccinationCertificateId)
-                }
-            )
-        )
+    private fun showDeleteVaccinationDialog(vaccinationCertificateId: String, position: Int?) {
+        MaterialAlertDialogBuilder(requireContext()).apply {
+            setTitle(R.string.vaccination_list_deletion_dialog_title)
+            setMessage(R.string.vaccination_list_deletion_dialog_message)
+            setPositiveButton(R.string.vaccination_list_deletion_dialog_positive_button) { _, _ ->
+                viewModel.deleteVaccination(vaccinationCertificateId)
+            }
+            setNegativeButton(R.string.vaccination_list_deletion_dialog_negative_button) { _, _ ->
+                position?.let { vaccinationListAdapter.notifyItemChanged(it) }
+            }
+            setOnCancelListener {
+                position?.let { vaccinationListAdapter.notifyItemChanged(it) }
+            }
+        }.show()
     }
 
     companion object {
